Refactor response repos
