+++ app/index.js
@@ -2,7 +2,8 @@ require('./insights').setup()
 require('log-timestamp')
 const messaging = require('./messaging')
 const createServer = require('./server')
-const { initialiseTables, initialiseContainer } = require('./storage/repos/responses')
+const { initialiseTables } = require('./storage/repos/responses')
+const { initialiseContainer } = require('./storage/repos/finalised-responses')
 
 const init = async () => {
   const server = await createServer()
@@ -12,7 +13,7 @@ const init = async () => {
     await initialiseTables()
     await initialiseContainer()
   }
-  
+
   console.log('Server running on %s', server.info.uri)
   await messaging.start()
 }
